Move password hashing methods to Password class removing PasswordArgon2 class
